Somehow got it working
