+++ App.js
@@ -1,13 +1,22 @@
 import { StatusBar } from 'expo-status-bar';
 import { StyleSheet, Text, View } from 'react-native';
-import MainMenu from "./src/app/(mainmenu)/mainmenu";
-import AuthScreen from "./src/app/(auth)/authscreen";
+import {NavigationContainer} from '@react-navigation/native';
+import {createNativeStackNavigator} from '@react-navigation/native-stack';
+import { createBottomTabNavigator } from '@react-navigation/bottom-tabs';
+import MainMenu from './src/app/mainmenu';
+import AuthScreen from './src/app/authscreen';
+
+const Tab = createBottomTabNavigator();
+
 
 export default function App() {
   return (
-    <View style={styles.container}>
-      <MainMenu />
-    </View>
+    <NavigationContainer>
+      <Tab.Navigator initialRouteName="MainMenu">
+        <Tab.Screen name='MainMenu' component={MainMenu} />
+        <Tab.Screen name='AuthScreen' component={AuthScreen} />
+      </Tab.Navigator>
+    </NavigationContainer>
   );
 }
 
@@ -15,6 +24,11 @@ const styles = StyleSheet.create({
   container: {
     flex: 1,
     backgroundColor: '#fff',
+    alignItems: 'center',
+    justifyContent: 'center',
+  },
+  text:{
+    color: "blue",
+    fontWeight: "bold",
   },
 });
-
