+++ README.md
@@ -1,1 +1,1 @@
-# groupproject-frontend-group-2
+# groupproject-frontend-group-3
