+++ app.json
@@ -1,7 +1,7 @@
 {
   "expo": {
-    "name": "QuantumCheckersIOS",
-    "slug": "QuantumCheckersIOS",
+    "name": "groupproject-frontend-group-3",
+    "slug": "groupproject-frontend-group-3",
     "version": "1.0.0",
     "orientation": "portrait",
     "icon": "./assets/icon.png",
