+++ package.json
@@ -1,5 +1,5 @@
 {
-  "name": "quantumcheckersios",
+  "name": "groupproject-frontend-group-3",
   "version": "1.0.0",
   "main": "node_modules/expo/AppEntry.js",
   "scripts": {
@@ -9,10 +9,16 @@
     "web": "expo start --web"
   },
   "dependencies": {
+    "@expo/vector-icons": "^14.0.0",
+    "@react-native-community/datetimepicker": "^7.6.4",
+    "@types/react": "~18.2.45",
     "expo": "~50.0.11",
+    "expo-router": "^3.4.8",
     "expo-status-bar": "~1.11.1",
-    "react": "18.2.0",
-    "react-native": "0.73.4"
+    "react": "^18.2.0",
+    "react-native": "0.73.4",
+    "react-time-picker": "^6.6.0",
+    "typescript": "^5.3.0"
   },
   "devDependencies": {
     "@babel/core": "^7.20.0"
